Fix testing errors; add unit test
Fix the ``load_spectrograph()`` location that I missed for updating
the allowed extensions.  Coadd1d files do not incude the [rdx] section,
so can't reference that.

Added unit test for the ``load_spectrograph()`` function that tests all
three methods for instantiating a Spectrograph class and also tests the
new "pypeit_fits" boolean flag at the heart of this PR (#2013).

	modified:   pypeit/scripts/coadd_1dspec.py
	modified:   pypeit/spectrographs/util.py
	modified:   pypeit/tests/test_spectrographs.py